# openapi/hinaa-api.yaml
openapi: 3.1.0
info:
  title: HINAA API
  version: 0.3.0-phase3
  description: Phase 3 live streaming plus preserved Phase 2 REST/mock fallbacks and the Phase 0 contract. Tools, BYOK execution, payments and continuous vision are not enabled.
...
servers:
  - url: https://api.example.invalid/v1
    description: Placeholder only
security:
  - sessionCookie: []
paths:
...
  /sessions/{sessionId}:
    delete:
      operationId: completeSession
      parameters: [{$ref: '#/components/parameters/SessionId'}, {$ref: '#/components/parameters/Csrf'}]
      responses:
        '204': {description: Completed}
        '404': {$ref: '#/components/responses/Error'}
...
  /conversations/turns:stream:
    post:
      security: []
      operationId: streamPhase2Turn
      description: Returns validated application events as newline-delimited JSON. The complete TurnPlan is validated before text deltas are emitted.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Phase2TurnRequest'}}}}
      responses:
        '200': {description: NDJSON thinking, text.delta, plan, usage or typed error events, content: {application/x-ndjson: {schema: {type: string}}}}
        '422': {$ref: '#/components/responses/Error'}
  /speech/transcriptions:
    post:
      security: []
      operationId: transcribePhase2Recording
      description: Accepts at most 20 seconds / 4 MiB of 16 kHz 16-bit mono PCM WAV; request audio is not retained.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [audio]
              properties:
                audio: {type: string, contentMediaType: audio/wav}
                language: {type: string, default: ne-NP}
                provider_mode: {enum: [mock, real], default: mock}
      responses:
        '200': {description: Final record-then-process transcript, content: {application/json: {schema: {$ref: '#/components/schemas/Phase2Transcript'}}}}
        '413': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /speech/synthesis:
    post:
      security: []
      operationId: synthesizePhase2Speech
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Phase2SpeechRequest'}}}}
      responses:
        '200': {description: 16 kHz 16-bit mono PCM WAV, content: {audio/wav: {schema: {type: string, contentMediaType: audio/wav}}}}
        '422': {$ref: '#/components/responses/Error'}
...
  /privacy/export:
    post:
      operationId: requestDataExport
      parameters: [{$ref: '#/components/parameters/Csrf'}]
      responses:
        '202': {description: Export queued}
  /privacy/account:
    delete:
      operationId: requestAccountDeletion
      parameters: [{$ref: '#/components/parameters/Csrf'}]
      responses:
        '202': {description: Account deletion queued}
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: __Host-hinaa_session}
  parameters:
    Csrf: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
...
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Error:
      description: Typed safe error
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, userActionRequired, correlationId]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string, maxLength: 500}
        retryable: {type: boolean}
        userActionRequired: {type: boolean}
        correlationId: {type: string, format: uuid}
    Phase2TurnRequest:
      type: object
      additionalProperties: false
      required: [sessionId, text, companionId, language, providerMode]
      properties:
        sessionId: {type: string, pattern: '^[A-Za-z0-9_-]+$', minLength: 1, maxLength: 80}
        text: {type: string, minLength: 1, maxLength: 8000}
        companionId: {enum: [hinaa, hiro]}
        language: {enum: [ne-NP, en-US, hi-IN, mixed]}
        providerMode: {enum: [mock, real]}
        personality:
          type: object
          additionalProperties: false
          description: Optional bounded expression controls; omitted fields use server defaults.
          properties:
            affection: {type: number, minimum: 0, maximum: 0.8}
            sass: {type: number, minimum: 0, maximum: 0.7}
            energy: {type: number, minimum: 0, maximum: 0.9}
            humor: {type: number, minimum: 0, maximum: 0.8}
            proactivity: {type: number, minimum: 0, maximum: 0.6}
    Phase2Transcript:
      type: object
      additionalProperties: false
      required: [text, language, provider, latencyMs]
      properties:
        text: {type: string, minLength: 1, maxLength: 8000}
        language: {type: string, maxLength: 20}
        provider: {type: string, maxLength: 100}
        latencyMs: {type: integer, minimum: 0}
    Phase2SpeechRequest:
      type: object
      additionalProperties: false
      required: [text, companionId, providerMode]
      properties:
        text: {type: string, minLength: 1, maxLength: 4000}
        companionId: {enum: [hinaa, hiro]}
        providerMode: {enum: [mock, real]}